[feat] refatoração 'major'. Quebrando codigo em partes mantendo a experiencia. Correção de namespaces (pode causar quebra de codigo para os consumidores)
