+++ BrasilAPI/BrasilAPI/Responses/BaseResponse.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Runtime.Serialization;
-using System.Text;
-
-namespace SDKBrasilAPI
+﻿namespace SDKBrasilAPI.Responses
 { 
     public abstract class BaseResponse
     {
