+++ BrasilAPI/BrasilAPI/Responses/CNPJResponse.cs
@@ -1,9 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Runtime.Serialization;
-using System.Text;
 
-namespace SDKBrasilAPI
+namespace SDKBrasilAPI.Responses
 {
     [DataContract]
     public class CNPJResponse : BaseResponse
