+++ BrasilAPI/BrasilAPI/Responses/FIPEResponse.cs
@@ -1,10 +1,7 @@
-﻿using SDKBrasilAPI;
-using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Runtime.Serialization;
-using System.Text;
 
-namespace SDKBrasilAPI
+namespace SDKBrasilAPI.Responses
 {
     public class FIPEMarcasResponse : BaseResponse
     {
