+++ BrasilAPI/BrasilAPI/Responses/FeriadosResponse.cs
@@ -1,9 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Runtime.Serialization;
-using System.Text;
 
-namespace SDKBrasilAPI
+namespace SDKBrasilAPI.Responses
 {
     public class FeriadosResponse : BaseResponse
     {
